Lab/Lab2/requirements.txt
requests
firebase-admin
google-cloud-firestore
google-cloud-storage
google-cloud-language
functions-framework==3.3.0
google-events==0.4.0
google-api-core==2.6.0
protobuf==4.22.1
cloudevents==1.9.0
pandas
google-api-core
grpcio
setuptools
